+++ tests/testthat/test-build-readme.R
@@ -1,4 +1,4 @@
-test_that("can build README in root directory", {
+test_that("can build README.Rmd in root directory", {
   skip_on_cran()
 
   pkg <- local_package_create()
@@ -14,7 +14,7 @@ test_that("can build README in root directory", {
   expect_false(file_exists(path(pkg, "README.html")))
 })
 
-test_that("can build README in inst/", {
+test_that("can build README.Rmd in inst/", {
   skip_on_cran()
 
   pkg <- local_package_create()
@@ -37,6 +37,47 @@ test_that("can build README in inst/", {
   expect_false(file_exists(path(pkg, "inst", "README.html")))
 })
 
+test_that("can build README.qmd in root directory", {
+  skip_on_cran()
+  skip_if_not_installed("quarto")
+  skip_if(!nzchar(Sys.which("quarto")), "quarto cli not available")
+
+  pkg <- local_package_create()
+  usethis::ui_silence(
+    usethis::with_project(
+      pkg,
+      use_readme_qmd(open = FALSE)
+    )
+  )
+
+  build_readme(pkg, quiet = TRUE)
+  expect_true(file_exists(path(pkg, "README.md")))
+})
+
+test_that("can build README.qmd in inst/", {
+  skip_on_cran()
+  skip_if_not_installed("quarto")
+  skip_if(!nzchar(Sys.which("quarto")), "quarto cli not available")
+
+  pkg <- local_package_create()
+  usethis::ui_silence(
+    usethis::with_project(
+      pkg,
+      use_readme_qmd(open = FALSE)
+    )
+  )
+  dir_create(pkg, "inst")
+  file_move(
+    path(pkg, "README.qmd"),
+    path(pkg, "inst", "README.qmd")
+  )
+
+  build_readme(pkg, quiet = TRUE)
+  expect_true(file_exists(path(pkg, "inst", "README.md")))
+  expect_false(file_exists(path(pkg, "README.qmd")))
+  expect_false(file_exists(path(pkg, "README.md")))
+})
+
 test_that("useful errors if too few or too many", {
   pkg <- local_package_create()
   expect_snapshot(build_readme(pkg), error = TRUE)
@@ -49,6 +90,13 @@ test_that("useful errors if too few or too many", {
   )
   dir_create(pkg, "inst")
   file_copy(path(pkg, "README.Rmd"), path(pkg, "inst", "README.Rmd"))
+  expect_snapshot(build_readme(pkg), error = TRUE)
+})
+
+test_that("errors if both README.qmd and README.Rmd exist", {
+  pkg <- local_package_create()
+  file_create(path(pkg, "README.Rmd"))
+  file_create(path(pkg, "README.qmd"))
   expect_snapshot(build_readme(pkg), error = TRUE)
 })